+++ CHANGELOG.md
@@ -11,6 +11,20 @@ tag versions. The Bond compiler (`gbc`) and
 different versioning scheme, following the Haskell community's
 [package versioning policy](https://wiki.haskell.org/Package_versioning_policy).
 
+## Unreleased #
+
+* `gbc` & compiler library: TBD
+* IDL core version: TBD
+* IDL comm version: TBD
+* C++ version: TBD
+* C# NuGet version: TBD
+* C# Comm NuGet version: minor version bump needed
+
+### C# Comm ###
+* `EpoxyTransport` can be configured to enable TCP keep-alive to help detect
+  dead connections. See `EpoxyTransportBuilder.SetKeepAliveTimes` for
+  details.
+
 ## 5.0.0: 2016-09-12 #
 
 * `gbc` & compiler library: 0.6.0.0